requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.0
Flask==3.0.3
-e git+https://github.com/pallets-eco/flask-admin.git@1a18428e1408dd2328dbe5441730bd3ab290fa1f#egg=Flask_Admin
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
